Refactor: 상세 액션 UI 분리 #157
