+++ lib/features/plant/presentation/widgets/plant_detail_menu_button.dart
@@ -1,0 +1,79 @@
+import 'package:commonplant_frontend/core/assets/app_icon_assets.dart';
+import 'package:commonplant_frontend/core/theme/app_colors.dart';
+import 'package:commonplant_frontend/core/theme/app_sizes.dart';
+import 'package:commonplant_frontend/core/theme/app_spacing.dart';
+import 'package:commonplant_frontend/shared/widgets/common_edit_delete_popup.dart';
+import 'package:commonplant_frontend/shared/widgets/common_svg_icon.dart';
+import 'package:flutter/material.dart';
+
+class PlantDetailMenuButton extends StatelessWidget {
+  const PlantDetailMenuButton({
+    super.key,
+    required this.onEdit,
+    required this.onDelete,
+  });
+
+  final VoidCallback onEdit;
+  final VoidCallback onDelete;
+
+  @override
+  Widget build(BuildContext context) {
+    return Semantics(
+      button: true,
+      label: '식물 상세 메뉴',
+      child: ExcludeSemantics(
+        child: IconButton(
+          tooltip: '식물 상세 메뉴',
+          onPressed: () => _showMenu(context),
+          padding: EdgeInsets.zero,
+          constraints: const BoxConstraints.tightFor(
+            width: AppSizes.navigationBarSideWidth,
+            height: AppSizes.navigationBarHeight,
+          ),
+          icon: const CommonSvgIcon(
+            AppIconAssets.shape,
+            width: 4,
+            height: 20,
+            color: AppColors.textStrong,
+          ),
+        ),
+      ),
+    );
+  }
+
+  void _showMenu(BuildContext context) {
+    showGeneralDialog<void>(
+      context: context,
+      barrierDismissible: true,
+      barrierLabel: '식물 상세 메뉴 닫기',
+      barrierColor: Colors.black.withValues(alpha: 0.4),
+      transitionDuration: Duration.zero,
+      pageBuilder: (dialogContext, _, _) {
+        return Material(
+          type: MaterialType.transparency,
+          child: Stack(
+            children: [
+              Positioned(
+                top:
+                    MediaQuery.paddingOf(dialogContext).top +
+                    AppSizes.navigationBarHeight +
+                    AppSpacing.x8,
+                right: AppSpacing.x20,
+                child: CommonEditDeletePopup(
+                  onEdit: () {
+                    Navigator.of(dialogContext).pop();
+                    onEdit();
+                  },
+                  onDelete: () {
+                    Navigator.of(dialogContext).pop();
+                    onDelete();
+                  },
+                ),
+              ),
+            ],
+          ),
+        );
+      },
+    );
+  }
+}
